Refactored skins data structure
Updated the structure used to store the race and skin info so that
they're all referenced by a common key (a constant representing the
offset). This also allows the code to loop through races, instead of
using three slightly different variations of the same line of code in
several places.